JWT token payload for OAuth2
Added option to generate a JWT token to use in the form values payload. If JWT settings are provided, the "grant_type", "username", "password" and "refresh_token" keys are not added, as the JWT token contains all the relevant data to authenticate. Note: The current implementation is based on how Adobe uses it.