Minimize chance for failing to update WTS
In some cases it happened that during the deletion of a DLL the auto updater had no access to the file.

When the WTS has been stopped the updater waits for 1 minute before continuing to give the system time to release all resources.

Deletion is tried 5 times with a delay.

If it still fails an extra email notification has been added where this was missing to notify the developers.