Updated scripts service
It now uses the GCL's JavaScript interpreter instead of Lua. Lua wasn't as versatile as I had hoped and had incredibly bad JSON support. Since JSON originates from JavaScript it was a much better choice to work with.

For now, scripts will require a result set to work with and get the entire result set supplied as a global object called "WtsResultSet". The script needs to return the WtsResultSet as the end of the script, otherwise an error occurs.