Add support to process mutliple entities at once
Currently when using the document store read action only 1 entity is processed. Support has been added to process multiple in the same action so the table does not have to be read for each individual entity.

Furthermore support is added to not store the results in the MySQL database cache.