JavaScript interpreter (#325)
* Started work on script interpreter for Lua

* Updated scripts service

It now uses the GCL's JavaScript interpreter instead of Lua. Lua wasn't as versatile as I had hoped and had incredibly bad JSON support. Since JSON originates from JavaScript it was a much better choice to work with.

For now, scripts will require a result set to work with and get the entire result set supplied as a global object called "WtsResultSet". The script needs to return the WtsResultSet as the end of the script, otherwise an error occurs.

* Finalized script action

* Finalized JavaScript interpreter

Updated method name and added package reference for the JavaScript
interpreter (it's no longer part of the GCL base library).

---------

Co-authored-by: Hans <35447467+HansJuice@users.noreply.github.com>